Move rota de obtenção de apostas da liga.
