+++ controllers/bets.controller.js
@@ -11,18 +11,6 @@ const _getBetById = async (req, res) => {
         res.status(404).send({ message: 'Bet not found' });
 };
 
-const _getLeagueBets = async (req, res) => {
-    const betFilter = { 
-        leagueId: res.league._id.toString() 
-    };
-    if (req.query.userId) betFilter.userId = req.query.userId;
-    if (req.query.gpId) betFilter.gpId = req.query.gp;
-    const bets = await betService.findBet(betFilter);
-    return bets.length > 0 ?
-        res.status(200).send(bets) : 
-        res.status(404).send({ message: 'No bets found.' });
-};
-
 const _postNewBet = async (req, res) => {
     return res.status(200).send(await betService.createBet(req.body));
 };
@@ -35,7 +23,6 @@ const _patchBet = async (req, res) => {
 
 module.exports = {
     getBetById: _getBetById,
-    getLeagueBets: _getLeagueBets,
     postNewBet: _postNewBet,
     patchBet: _patchBet,
 };
