+++ routes/bets.router.js
@@ -3,7 +3,6 @@ const betValidator = require('../validators/bets.validator');
 const gpValidator = require('../validators/gps.validator');
 const betsController = require('../controllers/bets.controller');
 
-router.get('', betsController.getLeagueBets);
 router.get('/:id', betsController.getBetById);
 router.post(
     '/new',
